server/scripts/requirements.txt
python-dotenv==1.0.1
PyYAML==6.0.1
cfn-flip==1.3.0 